pyproject.toml
[project]
name = "pyrtlfm"
version = "0.1.0"
description = "FM demodulator for RTL-SDR with squelch and user callback"
requires-python = ">=3.10"
dependencies = [
    "ipykernel>=7.1.0",
    "scipy>=1.7",
    "matplotlib>=3.10.8",
    "numpy>=1.20",
    "pyrtlsdr>=0.2.93",
    "setuptools",
    "pandas>=2.3.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["pyrtlfm"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
